neon: Move bloc and result to it's own library
