+++ packages/neon/neon/lib/src/bloc/result.dart
