+++ packages/neon/neon/lib/src/bloc/result_builder.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/widgets.dart';
-import 'package:neon/src/utils/result.dart';
+import 'package:neon/src/bloc/result.dart';
 import 'package:rxdart/rxdart.dart';
 
 typedef ResultWidgetBuilder<T> = Widget Function(BuildContext context, Result<T> snapshot);
