+++ packages/neon/neon/lib/src/blocs/accounts.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:flutter/foundation.dart';
+import 'package:neon/src/bloc/bloc.dart';
 import 'package:neon/src/blocs/apps.dart';
 import 'package:neon/src/blocs/capabilities.dart';
 import 'package:neon/src/blocs/user_details.dart';
@@ -10,7 +11,6 @@ import 'package:neon/src/models/account.dart';
 import 'package:neon/src/platform/platform.dart';
 import 'package:neon/src/utils/account_options.dart';
 import 'package:neon/src/utils/app_implementation.dart';
-import 'package:neon/src/utils/bloc.dart';
 import 'package:neon/src/utils/global_options.dart';
 import 'package:neon/src/utils/request_manager.dart';
 import 'package:neon/src/utils/storage.dart';
