+++ packages/neon/neon/lib/src/blocs/first_launch.dart
@@ -1,6 +1,6 @@
 import 'dart:async';
 
-import 'package:neon/src/utils/bloc.dart';
+import 'package:neon/src/bloc/bloc.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:shared_preferences/shared_preferences.dart';
