+++ packages/neon/neon/lib/src/blocs/login.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
+import 'package:neon/src/bloc/bloc.dart';
 import 'package:neon/src/models/account.dart';
-import 'package:neon/src/utils/bloc.dart';
 import 'package:nextcloud/nextcloud.dart';
 import 'package:package_info_plus/package_info_plus.dart';
 import 'package:rxdart/rxdart.dart';
