+++ packages/neon/neon/lib/src/blocs/next_push.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
+import 'package:neon/src/bloc/bloc.dart';
 import 'package:neon/src/blocs/accounts.dart';
 import 'package:neon/src/models/account.dart';
-import 'package:neon/src/utils/bloc.dart';
 import 'package:neon/src/utils/global_options.dart';
 import 'package:rxdart/rxdart.dart';
