+++ packages/neon/neon/lib/src/blocs/timer.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 import 'dart:ui';
 
 import 'package:meta/meta.dart';
-import 'package:neon/src/utils/bloc.dart';
+import 'package:neon/src/bloc/bloc.dart';
 
 abstract class TimerBlocEvents {
   /// Register a [callback] that will be called periodically.
