+++ packages/neon/neon/lib/src/blocs/user_details.dart
@@ -1,9 +1,9 @@
 import 'dart:async';
 
+import 'package:neon/src/bloc/bloc.dart';
+import 'package:neon/src/bloc/result.dart';
 import 'package:neon/src/models/account.dart';
-import 'package:neon/src/utils/bloc.dart';
 import 'package:neon/src/utils/request_manager.dart';
-import 'package:neon/src/utils/result.dart';
 import 'package:nextcloud/nextcloud.dart';
 import 'package:rxdart/rxdart.dart';
