+++ packages/neon/neon/lib/src/blocs/user_statuses.dart
@@ -1,11 +1,11 @@
 import 'dart:async';
 
 import 'package:flutter/foundation.dart';
+import 'package:neon/src/bloc/bloc.dart';
+import 'package:neon/src/bloc/result.dart';
 import 'package:neon/src/blocs/timer.dart';
 import 'package:neon/src/models/account.dart';
 import 'package:neon/src/platform/platform.dart';
-import 'package:neon/src/utils/bloc.dart';
-import 'package:neon/src/utils/result.dart';
 import 'package:nextcloud/nextcloud.dart';
 import 'package:rxdart/rxdart.dart';
 import 'package:window_manager/window_manager.dart';
