+++ packages/neon/neon/lib/src/utils/app_implementation.dart
@@ -2,10 +2,10 @@ import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:neon/l10n/localizations.dart';
+import 'package:neon/src/bloc/bloc.dart';
 import 'package:neon/src/blocs/accounts.dart';
 import 'package:neon/src/models/account.dart';
 import 'package:neon/src/platform/platform.dart';
-import 'package:neon/src/utils/bloc.dart';
 import 'package:neon/src/utils/nextcloud_app_specific_options.dart';
 import 'package:neon/src/utils/request_manager.dart';
 import 'package:neon/src/utils/storage.dart';
