+++ packages/neon/neon/lib/src/widgets/account_tile.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:neon/src/bloc/result_builder.dart';
 import 'package:neon/src/blocs/accounts.dart';
 import 'package:neon/src/models/account.dart';
 import 'package:neon/src/widgets/exception.dart';
 import 'package:neon/src/widgets/linear_progress_indicator.dart';
-import 'package:neon/src/widgets/result_builder.dart';
 import 'package:neon/src/widgets/user_avatar.dart';
 import 'package:nextcloud/nextcloud.dart';
 import 'package:provider/provider.dart';
