+++ packages/neon/neon/lib/utils.dart
@@ -1,5 +1,4 @@
 export 'package:neon/src/utils/app_implementation.dart';
-export 'package:neon/src/utils/bloc.dart';
 export 'package:neon/src/utils/branding.dart';
 export 'package:neon/src/utils/confirmation_dialog.dart';
 export 'package:neon/src/utils/exceptions.dart';
@@ -7,7 +6,6 @@ export 'package:neon/src/utils/hex_color.dart';
 export 'package:neon/src/utils/nextcloud_app_specific_options.dart';
 export 'package:neon/src/utils/rename_dialog.dart';
 export 'package:neon/src/utils/request_manager.dart';
-export 'package:neon/src/utils/result.dart';
 export 'package:neon/src/utils/sort_box_builder.dart';
 export 'package:neon/src/utils/sort_box_order_option_values.dart';
 export 'package:neon/src/utils/storage.dart';
