+++ packages/neon/neon/lib/widgets.dart
@@ -5,4 +5,3 @@ export 'package:neon/src/widgets/image_wrapper.dart';
 export 'package:neon/src/widgets/linear_progress_indicator.dart';
 export 'package:neon/src/widgets/list_view.dart';
 export 'package:neon/src/widgets/relative_time.dart';
-export 'package:neon/src/widgets/result_builder.dart';
